pyproject.toml
[project]
name = "simport"
version = "0.1.0"
description = "Tool to import songs to ultrastar/vibes using base lyric timings from lrclib.net and demucs/whisperx word level refinement."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "requests>=2.33.1",
    "whisperx>=3.8.5",
    "click>=8.0.0",
    "demucs>=4.0.1",
    "soundfile>=0.13.1",
    "yt-dlp>=2025.1.26",
    "google-genai>=1.75.0",
    "python-socketio>=5.16.1",
    "websocket-client>=1.9.0",
]

[project.scripts]
simport = "simport.cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
